Add Optional Memory Limit Configuration to Sandbox
Implemented an optional memory limit for the sandbox environment in the `compose.yml` file. This allows users to set a maximum memory usage for the sandbox by configuring the `SANDBOX_MEMORY_LIMIT` environment variable. The default limit is set to 100GB. This change addresses the issue of excessive memory consumption during data analysis flows, enhancing system stability for resource-intensive tasks.